Add Consume v3
In Consume v3 the node sends the records in batches to the client. After
sending a batch the node waits for the client to send a Continue message,
then the node sends the next batch. Suspending and resuming the request
happens only inside the client, the iteration over the records in the most
recently received batch is suspended and resumed. This allows for
guaranteed instantaneous request suspension.