Disable Nagle's algorithm in client and remove explicit flushing
As per dmqproto#48, TCP_CORK based flushing doesn't work in general,
however it's also no longer necessary, since we're doing explicit
buffering via record batching in Consume, and for all the control
messages, we're anyway flushing as soon as we send it. This makes
TCP_NODELAY fit for this purpose: as soon as we write the control
message it will be sent, and the batch will be sent as soon as sendmsg
returns, so there's no need for explicit flushing.

Fixes #48