Fix race condition in the `Producer` dmq test, causing the test to hang
It is possible that the `dmqtest.DmqClient.Producer.producer` callback is
called immediately when the `Produce` request is assigned. This causes
`waitNextEvent` to stall the application.

The error sequence was:

- Assign the `Produce` request. This may call the
  `DmqClient.Producer.producer` callback, in which case the error happens.
- Call `waitNextEvent`, which suspends the task, to be resumed by
  `DmqClient.Producer.producer`. This callback has already been called so
  the task is never resumed.

Fix this error by making `waitNextEvent` not suspend the task if
`producers.length` is not 0, which means that
`DmqClient.Producer.producer` has already been called so there is no need
to wait.