Apply d1to2fix (only `this` -> `(&this`))
Note: The refactors in this commit introduce compilation errors,
which should be resolved in a few successive commits.

The changes introduced by d1to2fix in this case have been mostly
reverted, wherever `(&this)` (changed from `this`) added an unnecessary
indirection.

The only case remaining in this commit is the one where the indirection
is specifically needed.