Remove unnecessary parser field from RequestHandler class
This class instance is already accessible through the `ed` field using
`ed.message_parser`, which is how it is used in all other d**protos.
The use of a separate `parser` field looks like a slightly mispurposed
attempt at DRY, which as a by-product relies on an implicitly public
import of `MessageParser` inside the `EventDispatcher` class.

Such implicitly public imports will break as of version 2.087.0 of the D
frontend, so it seems better to just drop this usage and use the event
dispatcher instance directly.

An alternative would be to have defined a `parser` method which returns
`ed.message_parser` (we can use `auto` return type to avoid any need to
rely on implicit imports), but this seems a bit overkill, and both this
and the original class field just obfuscate ownership of the parser.